Create Custom_Exception_1.java
Topic : Custom Exception(Exception Handling)

Write a Program to take care of Number Format Exception if the user enters values other than integer for calculating the average marks of 2 students. The name of the students and marks in 3 subjects are taken from the user while executing the program.
In the same Program write your own Exception classes to take care of Negative values and values out of range (i.e. other than in the range of 0-100).