Create Custom_Exception_2.java
Topic: Custom Exception(Exception Handling

Write a program to accept age of a person and ensure that the age entered is >=18 and < 60 else display invalid age is input .
Display proper error messages. 

The program must exit gracefully after displaying the error message in case the arguments passed are not proper.